Splitter ut klasser for å få litt bedre struktur
